add evaluable functor: div
