+++ bootstrap.pl
@@ -36,6 +36,7 @@
 :-(op(400, yfx, *)).
 :-(op(400, yfx, /)).
 :-(op(400, yfx, //)).
+:-(op(400, yfx, div)).
 :-(op(400, yfx, rem)).
 :-(op(400, yfx, mod)).
 :-(op(400, yfx, <<)).
